Fix backup node lookup
The node lookup code added in change
I27201319f31cdc01605a3c5ae9ef4b4218e4a3f6
was slightly broken in that we call a method
with a keyword arguemnt which doesn't exist.

uuid versus node_uuid.

It happens, it is a quick fix!

Spotted on a metalsmith job:

[-] Agent is requesting to perform an explicit node cache update.
    This is to pickup any chanages in the cache before deployment.
[-] Failed to update node cache. Error lookup_node() got an
    unexpected keyword argument 'uuid'

Change-Id: I59ecec65707a2f03918b233f1925395ebe59b8c4