CI: Lower memory usage of VMs/Increase swap
The IPA images we presently build/use for centos are
now smaller than their peak size a few months ago.

As such, we should be able to reduce that memory footprint
and still execcute all of the necessary tests without any
issues.

Also increases the amount of swap to 8GB as the infra
default is now 1GB.

Change-Id: I5705f1f2504803a049ed88a0d8b1d6d0bbccf9e4