Pin version of ipa-builder when publishing image
The ironic-python-agent-build-image-tinyipa job from ipa-builder comes
from master branch and it will use that branch to build all the tinyipa
ramdisks.
For compatibility reasons with the base version of tinycore used to
build tinyipa, we need to pin the ipa-builder code to tag 2.2.0 as
ussuri CI is based on bionic and it won't support the new kernel used in
tinycore 11.x when we need to chroot to modify the image during CI jobs
like ironic-standalone.

Change-Id: I60682d7e1d83793fb9dc3df4398e8e6f17438f98