unit tests: Capture log output
Debugging unit test failures is very challenging because all of the log
output is mixed together. Capture the log messages and store them as
part of the test, so that only log messages associated with failures are
output, interleaved correctly with the results.

This does not capture logs sent to stdout by _early_log(), although that
would be a nice further improvement.

Change-Id: I01e409ea73ff444e16caae0078e81585c703d5a5