Train only - Fix py3 support for bootloader default config load
When the support to load the default bootloader configuration was
backported to stable train, I modified it to pass python2 unit tests
as well as ultimately attempt to be compatible.

It might have actually worked! What did not work for Python3 when the
UTF-16 codepoint delimiter file was encountered in a production case.
This passed CI originally just fine, however Ubuntu images do not have
the delimiter codepoint, where as RHEL images *do* have a UTF-16
delimiter at the start of the file.

On python3, with the python2 compatible code, the file was being opened
as ascii or UTF-8, which we know as the same rule applies for both. All
code points must be <0xFF. This resulted in the file open itself to fail,
as encoding support differs between Python2 and Python3.

What this patch does now is separately handle the case of if python2,
to use the new preview python3 io subsystem previewed in python2 to
open the file properly, and present the UTF-16 encoded results.

This issue was identified in RHOSP testing with python 3.6, with
partition images where the deployment would fail with reporting
that the ascii codec could not decode 0xFF which indicates that the
contents are being handled as something other than utf-16, and it
is hitting a constraint in string handling in Python where ascii and
utf-8 character code points are expected to always be <0xFF.

Story: 2009020
Task: 42742
rhbz#1977417

Change-Id: I08664c2b3900448633837e40098a0ff9f0bd65f7