CI: Zuul no longer respects queue param
Also disable use of swift tempurls as swift seems
to sporatically fail on accessing assets, and the
thing being tested here is not swift.

Disabling the redfish virutal media job, as it is failing
due to swift issues, and it is unlikely for them to be
fixed. Besides, there is no local override to bypass
swift on older versions.

Change-Id: I1e7a3b3c9ded13b10002bb47e98d4a7b486e1dd4