CI: Make swift/Vmedia job non-voting
It appears that swift is occassionally responding
to jobs with 404 when tempurls are used. Likely this is something
to do with time, but the underlying credenital can also be validated.

Since this works in newer branches, and no chanes can be attributed
to swift, there is likely some sort of intermediate library issue
where unable to pin down. In the interest of keeping a working CI,
for now, mark the job as non-voting until we're able to figure
something out.

Change-Id: I46e8e8800eb1b08caba8faab6247195ea646430b