Fix create configuration unit tests
The unit tests for create_configuration give different result if
ran on a bios or uefi booted machine because they get the
partition table type value based on the utils function
get_node_boot_mode.
Let's mock the boot_mode as we do in other tests to get an
independent result.

Change-Id: Ic0e7daea7ec4ce0806cd126c27166f84690c5d9e