Extend the lookup timeout to 600 seconds
Changes the default lookup timeout to be 600 seconds which
reduces the risk of lookup failing as a write operation
to the backing database is performed upon lookup thanks to
generation of an agent token.

Overall, this is fairly harmless since by default ramdisks
restart the agent if they were not able to successfully
start.

Change-Id: I35c64c0b4f9b3b607df1bc0c4c2a852aa3595cbd