Test coverage for efi_utils.get_boot_record
A step will be developed to delete all EFI entries of type HD. As part
of this get_boot_record will need to parse more of the output of
`efibootmgr -v`.

This change asserts the existing behaviour of get_boot_record, and the
test can evolve with the changes in get_boot_record.

Related-Bug: #2041901
Change-Id: I0c5ac4adc1044c528c27a4eaf580c619ceef47e0